Fix tests failing
